# api/openapi/ecommerce-api.openapi.yaml
openapi: 3.1.0
info:
  title: Scalable E-commerce Backend API
  version: 0.1.0
  description: >-
    Async, API-first FastAPI e-commerce backend. Hand-authored contract and
    design source of truth (Spectral-linted). Paths are filled in per phase.
  license:
    name: MIT
  contact:
    name: Scalable E-commerce Backend
    url: https://github.com/KRISHNATEJATATA/scalable-ecommerce-backend
servers:
  - url: http://localhost:8000
    description: Local development
tags:
  - name: health
    description: Liveness and readiness probes
  - name: metrics
    description: Prometheus metrics exposition
  - name: identity
    description: OIDC resource-server identity — the local user mirror and Keycloak admin management
  - name: catalog
    description: Product catalog — merchant product CRUD and the shopper listing
paths:
  /v1/products:
    get:
      tags: [catalog]
      operationId: listProducts
      summary: List products (keyset-paginated, filterable)
      description: >-
        Keyset-paginated listing of live (non-deleted) products. `sort` is a
        whitelisted field (`created_at`/`price`/`name`) with an optional leading
        `-` for descending; `category`/`merchant_id` filter the result.
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: sort
          in: query
          schema: { type: string, default: "-created_at" }
        - name: cursor
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - name: merchant_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: One keyset page of products.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPage"
        "400":
          description: Unknown sort/filter field or an undecodable cursor.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [catalog]
      operationId: createProduct
      summary: Create a product (merchant only)
      description: >-
        Creates a product owned by the authenticated merchant. `merchant_id` is
        never accepted from input — it is bound from the caller. Emits
        `ProductCreated` via the transactional outbox.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreate"
      responses:
        "201":
          description: The created product.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          description: Invalid request body (e.g. non-positive price).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /v1/products/{product_id}:
    parameters:
      - $ref: "#/components/parameters/ProductId"
    get:
      tags: [catalog]
      operationId: getProduct
      summary: Get one product
      description: Returns one live (non-deleted) product by id, or 404.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The product.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [catalog]
      operationId: updateProduct
      summary: Update a product (owner merchant only)
      description: >-
        Partial update of an owned product. A merchant mutating another
        merchant's product → 403 (admin bypasses ownership). Emits
        `ProductUpdated` via the outbox.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: The updated product.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [catalog]
      operationId: deleteProduct
      summary: Soft-delete a product (owner merchant only)
      description: >-
        Soft-deletes an owned product (cross-merchant → 403, admin bypasses).
        Emits `ProductDeleted` via the outbox.
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Product soft-deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/products/{product_id}/image:presign:
    parameters:
      - $ref: "#/components/parameters/ProductId"
    post:
      tags: [catalog]
      operationId: presignProductImage
      summary: Presign an image upload (owner merchant only)
      description: >-
        Issues a short-TTL presigned S3 POST for an owned product image, but only
        AFTER validating ownership + content-type + declared size — the endpoint is
        never an open uploader. The upload lands under a private `uploads/` key; an
        S3-event worker then sniffs the real bytes, re-encodes (stripping EXIF),
        writes thumbnails, and marks the product image `ready` only if it passes
        (spoofed/oversized → `failed`). Cross-merchant → 403, missing → 404.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImagePresignRequest"
      responses:
        "200":
          description: The presigned-POST envelope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImagePresignResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/me:
    get:
      tags: [identity]
      operationId: getMe
      summary: Get the authenticated caller
      description: >-
        Returns the caller's JIT-provisioned local `users` mirror. First
        authenticated request provisions the row; subsequent calls reuse it.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The caller's local user mirror.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Missing, invalid, expired, or malformed bearer token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "403":
          description: The local account is disabled.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /v1/admin/users:
    post:
      tags: [identity]
      operationId: createUser
      summary: Create a Keycloak user (admin only)
      description: >-
        Provisions a brand-new Keycloak account. Keycloak defaults new users to
        the `consumer` realm role — the app never accepts a role from input.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: The new account's Keycloak `sub`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateUserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/admin/users/{sub}/roles/merchant:
    post:
      tags: [identity]
      operationId: grantMerchantRole
      summary: Grant the merchant realm role (admin only)
      description: >-
        Adds the Keycloak `merchant` realm role to the given user. Idempotent:
        granting a role the user already has still returns `204`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserSub"
      responses:
        "204":
          description: Role granted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [identity]
      operationId: revokeMerchantRole
      summary: Revoke the merchant realm role (admin only)
      description: >-
        Removes the Keycloak `merchant` realm role from the given user.
        Idempotent: revoking a role the user doesn't have still returns `204`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserSub"
      responses:
        "204":
          description: Role revoked.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/admin/users/{sub}/disable:
    post:
      tags: [identity]
      operationId: disableUser
      summary: Disable a user (admin only)
      description: >-
        Disables the account in Keycloak and mirrors `is_active=false` locally
        (provisioning the local row first if the user never authenticated).
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserSub"
      responses:
        "204":
          description: Account disabled.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/internal/whoami:
    get:
      tags: [identity]
      operationId: internalWhoami
      summary: Machine-to-machine identity check (service role only)
      description: >-
        Exercised by relay/worker clients authenticating with a Keycloak
        `service` machine-account token; confirms the role gate resolves
        end-to-end with no DB hit.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The caller's `sub` and role.
          content:
            application/json:
              schema:
                type: object
                properties:
                  sub:
                    type: string
                  role:
                    type: string
                required: [sub, role]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/health:
    get:
      tags: [health]
      operationId: getHealth
      summary: Liveness probe
      description: Returns 200 while the process is alive. Performs no dependency checks.
      responses:
        "200":
          description: Process is alive.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                required: [status]
  /v1/ready:
    get:
      tags: [health]
      operationId: getReady
      summary: Readiness probe
      description: Pings Postgres and Valkey; 200 when both are reachable, 503 otherwise.
      responses:
        "200":
          description: All critical dependencies are reachable.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  checks:
                    type: object
                    additionalProperties:
                      type: boolean
                required: [status, checks]
        "503":
          description: One or more critical dependencies are unreachable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /metrics:
    get:
      tags: [metrics]
      operationId: getMetrics
      summary: Prometheus metrics
      description: Prometheus text exposition of the default registry.
      responses:
        "200":
          description: Metrics in Prometheus text exposition format.
          content:
            text/plain:
              schema:
                type: string
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak-issued RS256 access token, validated against the realm's JWKS.
  parameters:
    UserSub:
      name: sub
      in: path
      required: true
      description: The target user's Keycloak OIDC `sub`.
      schema:
        type: string
    ProductId:
      name: product_id
      in: path
      required: true
      description: The product's UUID.
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Missing, invalid, expired, or malformed bearer token.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Authenticated but not permitted (missing role / not owner).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: The requested resource does not exist (or was soft-deleted).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    BadRequest:
      description: The request was malformed or failed validation (e.g. unsupported image type or size).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    User:
      description: The local `identity.users` mirror (JIT-provisioned, no `role` field).
      type: object
      properties:
        id:
          type: string
          format: uuid
        oidc_sub:
          type: string
        email:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, oidc_sub, email, is_active, created_at, updated_at]
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
          minLength: 3
          maxLength: 320
        temporary_password:
          type: string
          minLength: 8
          maxLength: 128
      required: [email, temporary_password]
    CreateUserResponse:
      type: object
      properties:
        sub:
          type: string
      required: [sub]
    Product:
      description: A merchant's product listing.
      type: object
      properties:
        id:
          type: string
          format: uuid
        merchant_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: [string, "null"]
        category:
          type: [string, "null"]
        price:
          type: string
          description: Decimal (money) serialized as a string.
        image_key:
          type: [string, "null"]
          description: Processed public object key; set by the image worker only after the upload passes.
        image_status:
          type: string
          enum: [none, pending, ready, failed]
          description: Image pipeline state (worker-owned; never accepted from input).
        image_url:
          type: [string, "null"]
          description: Unsigned CDN URL for a `ready` image, else `null`.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, merchant_id, name, description, category, price, image_key, image_status, image_url, created_at, updated_at]
    ProductCreate:
      description: Merchant create payload. `merchant_id` is bound from the caller, never from input. `image_key`/`image_status` are worker-owned and not accepted here.
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: [string, "null"]
        category:
          type: [string, "null"]
          maxLength: 255
        price:
          type: string
          description: Positive decimal money value (> 0).
      required: [name, price]
    ProductUpdate:
      description: Partial merchant update — every field optional; unset fields are untouched. `image_key` is not updatable here (use the presign endpoint).
      type: object
      additionalProperties: false
      properties:
        name:
          type: [string, "null"]
          minLength: 1
          maxLength: 255
        description:
          type: [string, "null"]
        category:
          type: [string, "null"]
          maxLength: 255
        price:
          type: [string, "null"]
          description: Positive decimal money value (> 0).
    ImagePresignRequest:
      description: Merchant declares the intended upload; validated BEFORE a URL is issued.
      type: object
      additionalProperties: false
      properties:
        content_type:
          type: string
          description: Claimed image MIME type (image/jpeg, image/png, image/webp).
        content_length:
          type: integer
          minimum: 1
          description: Declared byte size; capped by the presigned-POST policy.
      required: [content_type, content_length]
    ImagePresignResponse:
      description: >-
        Presigned-POST envelope for a direct-to-S3 upload. `fields` already carry
        the content-type + `content-length-range` policy conditions, so an upload
        violating them is rejected by S3 itself.
      type: object
      properties:
        url:
          type: string
          description: S3 POST endpoint the client uploads to.
        fields:
          type: object
          additionalProperties:
            type: string
          description: Form fields (policy, signature, key, content-type) to POST alongside the file.
        key:
          type: string
          description: The private upload object key the file lands at.
        expires_in:
          type: integer
          description: Seconds until the presigned POST expires.
      required: [url, fields, key, expires_in]
    ProductPage:
      description: One keyset page of products.
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        next_cursor:
          type: [string, "null"]
          description: Opaque cursor for the next page; `null` on the last page.
      required: [items]
    Problem:
      description: RFC 9457 Problem Details (the one flat error shape).
      type: object
      properties:
        type:
          type: string
        status:
          type: integer
        title:
          type: string
        detail:
          type: string
        trace_id:
          type: string
        details:
          type: array
          items:
            type: object
      required: [type, status, title]